backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pyyaml>=6.0
python-dotenv>=1.0.0
watchfiles>=1.0.0
aiosqlite>=0.20.0
aiohttp>=3.11.0
asyncpg>=0.29.0
cachetools>=5.3.0
PyJWT[crypto]>=2.10.0
psycopg2-binary>=2.9.9
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-http>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
prometheus-client>=0.20.0
pathspec>=0.12.1
typer>=0.12.0
rich>=13.7.0
mcp>=1.8,<2
